notification/requirements.txt
python-socketio==5.10.0
python-engineio==4.8.0
aiohttp==3.9.1
python-dotenv==1.0.0